Add filters to posts_controller
Make it so only signed in users are allowed to make posts and see who
the author of every post is. In accordance with the ability to sign in
and out, the posts#index view changes dynamically depending on if the
user is signed in or not, and who they are.